Updated graphics/shotwell to 0.11.5
Upstream changes:

0.11.5:
-------
This release fixes an issue in
which Shotwell could crash when using the "Import from F-Spot" feature
for the subset of users who continued to experience this problem after
the 0.11.4 upgrade. We recommend that all users upgrade.

0.11.4;
-------
This release fixes two critical
issues present in all previous versions of Shotwell 0.11.x that could
cause Shotwell to crash when using the Import from F-Spot feature.
We recommend that all users upgrade.

0.11.3:
-------
This releases fixes two critical bugs, including:
   * Shotwell could crash at the end of photo imports where one or
more files failed to import correctly
   * Showell crashed when a new tag containing a slash (/) character
was created by context-clicking on the Tags item in the sidebar and
choosing  New
and improves error reporting in the publishing system.

0.11.2:
-------
  * Improved stability working with hierarchical tags
  * Importing hierarchical tags from F-Spot doesn't generate duplicate
top-level tags
  * Fixed "server redirect contained no session key" errors in the
Facebook Connector
  * Corrected problems with item counts over mixed media