Import alacconvert-0.0.20111026 as audio/alacconvert from wip/alacconvert
The Apple Lossless Audio Codec (ALAC) is an audio codec developed
by Apple and supported on iPhone, iPad, most iPods, Mac and iTunes.
ALAC is a data compression method which reduces the size of audio
files with no loss of information.  A decoded ALAC stream is
bit-for-bit identical to the original uncompressed audio file.

This is Apple Lossless Audio Codec encode and decode command.