benchmarks/iozone: Really add DragonFly support
It appears that the authors of Iozone tried to add DragonFly support without
actually building it.  The configuration was broken, and the iozone.c
was uncompilable.  With the two modified patches, it will compile now, and a
quick check using "iozone -a" seemed to work as expected.
No PKGREVISION bump necessary.