lang/gcc44: Mark NOT-FOR-DRAGONFLY
This package has never built on DragonFly, but it really is not needed as
the base compiler is gcc 4.4.7.  The sole package (databases/libcassandra)
that required lang/gcc44 was just changed to remove this requirement when
built on DragonFly.  This compiler is not worth the effort to fix for
DragonFly.