math/pari: Fix for binutils 2.22
The ld linker from binutils 2.22 has --no-copy-dt-needed-entries enabled by
default, which is exactly opposite of earlier versions of the linker.
Previously ld would recursively search libraries linked in by the object's
primary libraries during symbol resolution.  Now linker wants all libraries
explicit requested, otherwise one gets a message such as this:

ld: <object file>: undefined reference to symbol '<symbol>'
ld: note: '<symbol>' is defined in DSO <shared library> so try adding it
    to the linker command line.

If this message is encountered, the two options are either to explicitly
add it to the linker command as suggested, or use the
--no-copy-dt-needed-entries switch to return the linker to its previous
behavior.  In the case of math/pari, libstdc++ was added to the library
list for the plotfltk.o object.  This explicit link is required by DragonFly
which is migrated to binutils 2.22 as its default version.