This package's upstream distfile has hit on the wonderful idea of shipping symlinks into /usr/share/automake-1.11 instead of copies of various semi-standard gnuish files, including some scripts needed by the build.
Remove the patch that causes the build to try to run things direct
from ${PREFIX}/share/automake-1.11, as that's not the way things are
supposed to be done in a buildlink universe. However, adding a bl3.mk
to automake to allow using this stuff is itself wrong.

So instead, work around this mess by adding a post-extract hack that
deletes the symlinks and replaces them with the actual files from
automake.

Also, add a build dependence on automake so the files are actually
there when we try to reference them. (hi reinoud!)

No revbump needed as the package didn't build.

Grumble.