print/jcabc2ps: Remove trace of cat pages
Apparently references to man/cat* on the PLIST are now being filtered out
by PKG_DEVELOPER=yes.  Now there is an error that cat pages are present
without being on the PLIST when actually they are.

So remove the installation of the jcabc2ps cat page despite there being no
man page version.