Update to 2.2:
2.2:
The X Input Extension protocol version 2.2 is now available.

The main feature added in this version is support for multitouch devices
and the ability for clients to register for and receive touch events.

Multitouch support in XI 2.2 aims to
- support a dynamic number of simultaneous touch points,
- support devices that are both multitouch and traditional pointer devices,
- allow touchpoints to be either grouped together or handled separately,
- be backwards-compatible to pre-XI 2.2 clients through emulation of XI 2.x/XI 1.x and core
  pointer events.

For a description of the new features see the following posts:
- http://who-t.blogspot.com.au/2011/12/multitouch-in-x-getting-events.html
- http://who-t.blogspot.com.au/2011/12/multitouch-in-x-pointer-emulation.html
- http://who-t.blogspot.com.au/2012/01/multitouch-in-x-touch-grab-handling.html
- http://who-t.blogspot.com.au/2012/02/multitouch-in-x-multitouch-touchpads.html

Many thanks to all who have contributed to this protocol.

2.1:
We haven't had any change requests to the 2.1 protocol changes and the 2.2
changes are about to be done soon too. Time for a release.

inputproto contains the protocol specification and header files for the
X Input Extension. This release introduces two new features:
- smooth scrolling support allows devices to send scroll events through
  valuator information instead of the traditional button 4-7 clicks.
- new raw event behaviour allows clients to listen to raw events from a
  device even if that device is currently grabbed

For a description of the changes, see
http://who-t.blogspot.com/2011/09/whats-new-in-xi-21-smooth-scrolling.html
http://who-t.blogspot.com/2011/09/whats-new-in-xi-21-raw-events.html
http://who-t.blogspot.com/2011/09/whats-new-in-xi-21-xi2-defines.html
http://who-t.blogspot.com/2011/09/whats-new-in-xi-21-versioning.html

Other changes include packaging fixes and miscellaneous fixes.