emulators/simh: Prevent -lto switch on DragonFly
This package starting failing this week.  It was due to an change in the
system binutils which turned on link time optimization support.  This
was intended for external compilers.

The package picked up that the linker was capable of supporting -flto, but
then just assumed the compiler also was!  "4.4.7" was added to the list of
compilers to exclude.  Earlier versions are not necessary because the
-flto switch would be suppressed anyway due to binutils not supporting it.