Update to Asterisk 1.8.15.0; this is a general bug fix release:
The release of Asterisk 1.8.15.0 resolves several issues reported by the
community and would have not been possible without your participation.
Thank you!

The following is a sample of the issues resolved in this release:

* --- Fix deadlock potential with ast_set_hangupsource() calls.

* --- Fix request routing issue when outboundproxy is used.

* --- Make the address family filter specific to the transport.

* --- Fix NULL pointer segfault in ast_sockaddr_parse()

* --- Do not perform install on existing directories

For a full list of changes in this release, please see the ChangeLog:

http://downloads.asterisk.org/pub/telephony/asterisk/ChangeLog-1.8.15.0

Thank you for your continued support of Asterisk!