Speed up the algorithm to determine the files left to check.
This change modifies the algorithm used to keep track of the files that
have not yet been checksummed to use a simple loop instead of shell pattern
matching.

For packages with few distinfo entries, either way yields the same result
as the list of files to check is very short.  But for those packages with
hundreds of distinfo entries (vim, I'm looking at you), the difference is
huge.  In my old macppc machine, the checksum of vim used to take around
40 minutes and now it takes ~35 seconds.  The difference is also clearly
visible in my faster amd64 machine (although I haven't bothered to time it).