Add patch to work around upstream bug assuming beyond-POSIX behavior in sed.
In GNU sed, "\n" in the RHS of a substitution becomes a newline.
POSIX says that \n in the RHS is undefined.  BSD sed treats it as "n".
Because this is in scheme code, and scheme doesn't care about " " vs
"\n", just substitute a space.

Resolves failure of test cases; now "make test" passes.

(This change has been reported upstream and applied to the upstream VCS.)