Import gst-plugins1-good-1.0.3 as multimedia/gst-plugins1-good.
GStreamer is a library that allows the construction of graphs of
media-handling components, ranging from simple Ogg/Vorbis playback to
complex audio (mixing) and video (non-linear editing) processing.

Applications can take advantage of advances in codec and filter technology
transparently.  Developers can add new codecs and filters by writing a
simple plugin with a clean, generic interface.

GStreamer is released under the LGPL.

This package is part of the good GStreamer plugins; that is, those that are
considered to be stable and correctly coded.