Update to 5.0.2
Changelog:
Changes since 5.0.0
-------------------

Numeric constants encountered in mathematical expressions (but not other
contexts) can contain underscores as separators that will be ignored on
evaluation, as allowed in other scripting languages.  For example,
0xFFFF_FFFF, or 3.141_592_654.

"functions -T" turns on tracing for the specified function(s) only,
similar to "functions -t" except that tracing is turned off for any
functions called from the specified one(s) that don't also have the -t
or -T flag.

In file completion, the recursive-files style can be set to an array of
patterns to match against "$PWD/".  In any matched location, it is
possibly to complete files in arbitrarily deep subdirectories without
needing to type the directory prefix.  See example in the zshcompsys
manual.

The _user_expand completer now allows expansion functions in the
user-expand files to return a string in REPLY that will be used to name
the set of expansions returned.