Remove libboost_locale from the generic PLIST, it is handled in the OPSYS-specific files.
This is now the third time I have had to apply this identical fix,
please can package updaters be more careful in the future so I don't
have to do it a fourth time.  Alternatively, cookies are available to
anyone who comes up with a programatic fix for print-PLIST or whatever
is causing this to happen each time.