Explain xen version differences.
There are 5 versions of xen in pkgsrc, which is confusing.  Explain in
DESCR which version is in which package (xenkernel3 contains 3.1), and
which versions support PCI passthrough (only 3.1).  Explain which
versions support non-PAE (3.1) and PAE (3.3, 4.1, 4.2), because the
HOWTO is out of date and it's easy to end up with a non-working system
on a 3.1 to 3.3 update.  Cuation that 2.0 is beyond crufty.

This is a DESCR-only change (with PKGREVISION++ of course).

(ok during freeze agc@)