lang/gcc-aux: Keep original bootstrap for NetBSD/i386
The last change to the Makefile was created months ago, and at the time
it was intended that new bootstraps would be made for both x86_64 and
i386 platforms.  The latter bootstrap was never created, so the previous
commit broke the build on i386.

Until a new bootstrap compiler is created, continue to use the old one.
It will continue to build on gcc-based NetBSD, but I do not expect it to
build on NetBSD-6.99/i386 with clang and no gcc libraries.