Update ruby193-base and related packages to 1.9.3p545 (Ruby 1.9.3-p545).
pkgsrc chagnges:

* Use RUBY_SUFFIX instead of RUBY_VER for appropriate place.
* Detect NetBSD correctly in Gem::Platform.

Quote from release announce:

This is the last ordinal release of Ruby 1.9.3. It means that Ruby 1.9.3 goes
into the state of the security maintenance phase, and will never be released
unless any critical regressions or security issues are found.  This phase is
planned to be maintained for 1 year. Then, maintenance of Ruby 1.9.3 will be
ended at Feb. 24th, 2015.

This release includes many bugfixes. See tickets and ChangeLog for details.