pkgtools/pbulk: delete xcrun cache file before unprivileged su
On OS X, before running certain phases of a package build as the
unprivileged user via su, delete the xcrun cache file of the privileged
user to avoid "permission denied" errors in the package build logs when
the unprivileged xcrun program tries to overwrite the existing cache
file of the privileged user.

This is needed because, on OS X Yosemite with Command Line Tools
(10.10) for Xcode 6.4, the xcrun program creates an xcrun_db cache
file.  This cache file will be created when run as the privileged user.
Unfortunately, when xcrun is run as the unprivileged user via su, it
will try to overwrite the cache file of the privileged user and will
fail because it does not have the needed permissions.  This results
in the configure and build logs being littered with messages like the
following (the program name, "gm4", will vary):

  gm4: error: couldn't replace cache file '/var/tmp/xcrun_db' with newly
  written file '/var/tmp/xcrun_db-QqnQCjPl' (errno=Permission denied)