Update audio/beets to 1.4.3
pkgsrc changes:
 - Update MASTER_SITES to MASTER_SITE_PYPI
 - Update HOMEPAGE
 - Get rid of PYTHON_VERSIONS_INCOMPATIBLE
   (beets is now compatible with Python 3 too)
 - Misc cosmetic fixes and simplifications

Changes:
1.4.3 (January 9, 2017)
-----------------------

Happy new year! This new version includes a cornucopia of new features from
contributors, including new tags related to classical music and a new
:doc:`/plugins/absubmit` for performing acoustic analysis on your music. The
:doc:`/plugins/random` has a new mode that lets you generate time-limited
music---for example, you might generate a random playlist that lasts the
perfect length for your walk to work. We also access as many Web services as
possible over secure connections now---HTTPS everywhere!

1.4.2 (December 16, 2016)
-------------------------

This is just a little bug fix release. With 1.4.2, we're also confident enough
to recommend that anyone who's interested give Python 3 a try: bugs may still
lurk, but we've deemed things safe enough for broad adoption. If you can,
please install beets with ``pip3`` instead of ``pip2`` this time and let us
know how it goes!

1.4.1 (November 25, 2016)
-------------------------

Version 1.4 has **alpha-level** Python 3 support. Thanks to the heroic efforts
of :user:`jrobeson`, beets should run both under Python 2.7, as before, and
now under Python 3.4 and above. The support is still new: it undoubtedly
contains bugs, so it may replace all your music with Limp Bizkit---but if
you're brave and you have backups, please try installing on Python 3. Let us
know how it goes.


1.3.19 (June 25, 2016)
----------------------

This is primarily a bug fix release: it cleans up a couple of regressions that
appeared in the last version. But it also features the triumphant return of the
:doc:`/plugins/beatport` and a modernized :doc:`/plugins/bpd`.

It's also the first version where beets passes all its tests on Windows! May
this herald a new age of cross-platform reliability for beets.


1.3.18 (May 31, 2016)
---------------------

This update adds a new :doc:`/plugins/hook` that lets you integrate beets with
command-line tools and an :doc:`/plugins/export` that can dump data from the
beets database as JSON. You can also automatically translate lyrics using a
machine translation service.

The ``echonest`` plugin has been removed in this version because the API it
used is `shutting down`_. You might want to try the
:doc:`/plugins/acousticbrainz` instead.

.. _shutting down: https://developer.spotify.com/news-stories/2016/03/29/api-improvements-update/


1.3.17 (February 7, 2016)
-------------------------

This release introduces one new plugin to fetch audio information from the
`AcousticBrainz`_ project and another plugin to make it easier to submit your
handcrafted metadata back to MusicBrainz.
The importer also gained two oft-requested features: a way to skip the initial
search process by specifying an ID ahead of time, and a way to *manually*
provide metadata in the middle of the import process (via the
:doc:`/plugins/edit`).

Also, as of this release, the beets project has some new Internet homes! Our
new domain name is `beets.io`_, and we have a shiny new GitHub organization:
`beetbox`_.


1.3.16 (December 28, 2015)
--------------------------

The big news in this release is a new :doc:`interactive editor plugin
</plugins/edit>`. It's really nifty: you can now change your music's metadata
by making changes in a visual text editor, which can sometimes be far more
efficient than the built-in :ref:`modify-cmd` command. No more carefully
retyping the same artist name with slight capitalization changes.

This version also adds an oft-requested "not" operator to beets' queries, so
you can exclude music from any operation. It also brings friendlier formatting
(and querying!) of song durations.


1.3.15 (October 17, 2015)
-------------------------

This release adds a new plugin for checking file quality and a new source for
lyrics. The larger features are:

* A new :doc:`/plugins/badfiles` helps you scan for corruption in your music
  collection. Thanks to :user:`fxthomas`. :bug:`1568`
* :doc:`/plugins/lyrics`: You can now fetch lyrics from Genius.com.
  Thanks to :user:`sadatay`. :bug:`1626` :bug:`1639`
* :doc:`/plugins/zero`: The plugin can now use a "whitelist" policy as an
  alternative to the (default) "blacklist" mode. Thanks to :user:`adkow`.
  :bug:`1621` :bug:`1641`


1.3.14 (August 2, 2015)
-----------------------

This is mainly a bugfix release, but we also have a nifty new plugin for
`ipfs`_ and a bunch of new configuration options.


1.3.13 (April 24, 2015)
-----------------------

This is a tiny bug-fix release. It copes with a dependency upgrade that broke
beets.


1.3.12 (April 18, 2015)
-----------------------

This little update makes queries more powerful, sorts music more
intelligently, and removes a performance bottleneck. There's an experimental
new plugin for synchronizing metadata with music players.

Packagers should also note a new dependency in this version: the `Jellyfish`_
Python library makes our text comparisons (a big part of the auto-tagging
process) go much faster.


1.3.11 (April 5, 2015)
----------------------

In this release, we refactored the logging system to be more flexible and more
useful. There are more granular levels of verbosity, the output from plugins
should be more consistent, and several kinds of logging bugs should be
impossible in the future.

There are also two new plugins: one for filtering the files you import and an
evolved plugin for using album art as directory thumbnails in file managers.
There's a new source for album art, and the importer now records the source of
match data. This is a particularly huge release---there's lots more below.

There's one big change with this release: **Python 2.6 is no longer
supported**. You'll need Python 2.7. Please trust us when we say this let us
remove a surprising number of ugly hacks throughout the code.