Add ruby-proxifier package version 1.0.3.
Proxifier is a gem to force ruby to use a proxy.

This gem was created for 2 purposes.

First is to enable ruby programmers to use HTTP or SOCKS proxies
interchangeably when using TCPSockets. Either manually with
Proxifier::Proxy#open or by require "proxifier/env".

The second purpose is to use ruby code that doesn't user proxies for users
that have to use proxies.

The pruby and pirb executables are simple wrappers for their respective ruby
executables that support proxies from environment variables.